backend/requirements.txt
# Car Image AI - Backend Dependencies
python-dotenv>=1.0.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Gemini API (gemini-3.1-flash-image-preview) for car image processing
google-genai>=1.47.0

# Image handling
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
rembg>=2.0.50
onnxruntime>=1.16.0

# Nikon RAW (NEF) support
rawpy>=0.19.0
